+++ forui/lib/src/widgets/header/header.dart
@@ -1,8 +1,11 @@
+import 'dart:math';
 import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+import 'package:forui/src/foundation/rendering.dart';
 
 import 'package:meta/meta.dart';
 import 'package:sugar/collection.dart';